Prepare for WebGL Streaming on Website (#1746)
* chore: re-generate proto ts apis

* fix: message.getTeamId is not a function

* feat: stream page for THUAI7

* refactor: function call to webgl

* fix(contest): null error on RecordPage

* refactor(contest): text on RecordPage

* feat(playback): open file as PWA

* feat(playback): handle local files

* feat(playback)
