+++ src/app/ContestSite/RecordPage.tsx
@@ -69,8 +69,8 @@ const RecordPage: React.FC<ContestProps> = ({ mode, user }) => {
           ]
         : [],
       onFilter: (value, record) =>
-        record.contest_room_teams[0].contest_team.team_id === value ||
-        record.contest_room_teams[1].contest_team.team_id === value,
+        record.contest_room_teams[0]?.contest_team.team_id === value ||
+        record.contest_room_teams[1]?.contest_team.team_id === value,
       render: (text, record) => {
         return (
           <Text>
@@ -195,16 +195,7 @@ const RecordPage: React.FC<ContestProps> = ({ mode, user }) => {
       <Row>
         <Col span={2}></Col>
         <Col span={20}>
-          <Typography.Text mark>
-            历次对战的观战端口、结果和回放文件。直播及回放可使用
-            <Typography.Text code>wpf</Typography.Text>观看，直播观战URL
-            <Typography.Text code>IP: thuai6.eesast.com</Typography.Text>。
-          </Typography.Text>
-          <br />
-          <br />
-          <Typography.Text type="danger">
-            注：若观战端口长时间等待分配，说明本次对战请求失败，可能是服务器负载过高，也可能是队伍代码有误。
-          </Typography.Text>
+          <Typography.Text mark>在哪里跌倒，就在哪里爬起来！</Typography.Text>
         </Col>
       </Row>
       <br />
